% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipplotr-package.R
\docType{package}
\name{clipplotr-package}
\alias{clipplotr}
\alias{clipplotr-package}
\title{clipplotr: comparative visualisation of CLIP data}
\description{
Load CLIP crosslink, auxiliary (e.g. peak) and coverage tracks once with
\code{\link[=load_tracks]{load_tracks()}}, then plot any number of regions with \code{\link[=plot_region]{plot_region()}}.
The command-line tool (\code{./clipplotr}) is a thin wrapper around these
functions; see \code{\link[=run_cli]{run_cli()}}.
}
\author{
\strong{Maintainer}: Anob M. Chakrabarti \email{a.chakrabarti@ucl.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Charlotte Capitanchik [copyright holder]
}

Other contributors:
\itemize{
  \item Jernej Ule [copyright holder]
  \item Nicholas Luscombe [copyright holder]
}

}
\keyword{internal}
