% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{resolve_region}
\alias{resolve_region}
\title{Resolve a region of interest}
\usage{
resolve_region(region, annotation, no_ucsc = FALSE)
}
\arguments{
\item{region}{Coordinates as \code{chromosome:start:end:strand} (or
\code{chromosome:start-end:strand}), an Ensembl/GENCODE gene id (with or
without version suffix), or a gene name.}

\item{annotation}{Annotation from \code{\link[=load_annotation]{load_annotation()}}, used to look up gene
ids and names.}

\item{no_ucsc}{If \code{FALSE} (default), chromosome names in coordinates are
converted to UCSC style where possible.}
}
\value{
A GRanges of length 1 with a strand.
}
\description{
Resolve a region of interest
}
